Avaliações/Trabalhos/Bookshelf_API/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
PyMySQL==1.1.0
Werkzeug==2.3.7
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
